# service-managment/openapi/service-management.yaml
swagger: '2.0'
info:
  description: Servizi rest di gestione contributo per l'applicazione TariTefa
  version: 0.0.1
  title: Service Management REST API
host: 'localhost:8083'
basePath: /
tags:
  - name: service-management-controller
    description: Service Management Controller
paths:
  /service-management/service:
    post:
      tags:
        - service-management-controller
      summary: Salva il tributo
      description: Servizio REST per creare un tributo
      operationId: saveServiceUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: tributeServiceModel
          description: Modello del tributo
          required: true
          schema:
            $ref: '#/definitions/TributeServiceModel'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ServiceConfiguratedModel'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/service-management/service/info/{fiscalCode}':
    get:
      tags:
        - service-management-controller
      summary: Indica se è presente un tributo
      description: Servizio REST per ottenere l'informazione della presenza di un tributo configurato
      operationId: isServiceConfiguratedUsingGET
      produces:
        - '*/*'
      parameters:
        - name: fiscalCode
          in: path
          description: fiscalCode
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ServiceConfiguratedModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/service-management/service/{fiscalCode}':
    get:
      tags:
        - service-management-controller
      summary: Recupera i dati del tributo
      description: Servizio REST per recuperare i dati di un tributo
      operationId: getServiceUsingGET
      produces:
        - '*/*'
      parameters:
        - name: fiscalCode
          in: path
          description: fiscalCode
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/TributeServiceModel'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  InstallmentModel:
    type: object
    properties:
      dueDate:
        type: string
        format: date
      percentagePrimary:
        type: number
      percentageSecondary:
        type: number
    title: InstallmentModel
  ServiceConfiguratedModel:
    type: object
    properties:
      result:
        type: boolean
    title: ServiceConfiguratedModel
  TributeServiceModel:
    type: object
    properties:
      denomination:
        type: string
      dueDateUnique:
        type: string
        format: date
      fiscalCodePrimaryCreditor:
        type: string
      fiscalCodeSecondaryCreditor:
        type: string
      ibanPrimary:
        type: string
      ibanSecondary:
        type: string
      installments:
        type: array
        items:
          $ref: '#/definitions/InstallmentModel'
      percentageSecondary:
        type: number
    title: TributeServiceModel